Massive changes to fullscreen detection everywhere 	modified:   detection.sh 	modified:   focus.sh 	deleted:    focus2.sh 	modified:   fullscreen.sh 	modified:   position.sh 	modified:   pretty.sh 	modified:   setborder.sh 	modified:   tile.sh 	modified:   vroum.sh 	modified:   wgroups.sh 	modified:   winopen.sh

+++ pretty.sh
@@ -1,39 +1,69 @@
-#!/bin/sh
+#!/bin/mksh
 #
 # wildefyr - 2015 (c) wtfpl
-# Arrange windows in a *nix photogeneric way
-
-source fyrerc.sh
+# A script to put windows into a *nix photogeneric order
 
 ignore() {
-    cat $DETECT > $WLFILETEMP
+    if [ -e $DETECT ]; then
+        cat $DETECT > $WLFILETEMP
+    fi
     lsw >> $WLFILETEMP
 }
 
-prettyTile() {
-    ignore
-
+prettytile() {
     sort $WLFILETEMP | uniq -u | xargs wattr xi | sort -n | \
     awk '{print $2}' > $WLFILE
 
-    SW=$((SW / 2))
-    SH=$((SH / 2 + 200))
-    echo $SW $SH
+    SW=$((SW - 2*XGAP - BW))
+    SH=$((SH - TGAP - BGAP - BW))
 
-    for c in $(seq $COLS); do
-        if [ $c -eq 3 ]; then X=$((X + 1)); fi
-        wtp $X $Y $W $H $(head -n $c $WLFILE | tail -1)
-        X=$((X + W + IGAP))
+    COLS=2
+    W=$(((SW - (COLS - 1)*IGAP)/COLS))
+    H=$SH
+
+    COLSTEMP=$((COLS - 1))
+    cat $WLFILE | head -n $COLSTEMP > $WLFILETEMP
+
+    for c in $(seq $COLSTEMP); do
+        wtp $X $Y $W $H $(head -n $c $WLFILETEMP | tail -1)
+        X=$((X + W + IGAP - BW))
     done
 
+    cat $WLFILE | tail -n $((windowsOnscreen - COLSTEMP)) > $WLFILETEMP
+
+    ROWS=$(cat $WLFILETEMP | wc -l)
+    H=$(((SH - (ROWS - 1)*VGAP)/ROWS))
+    if [ $COLS -eq 3 ]; then
+        W=$((W + maxHorizontalWindows - 1))
+    fi
+
+    for c in $(seq $ROWS); do
+        if [ $c -ne 1 ]; then
+            if [ $((c % 2)) -eq 1 ]; then
+                H=$((H + 1))
+            fi
+        fi
+        wtp $X $Y $W $H $(head -n $c $WLFILETEMP | tail -1)
+        Y=$((Y + H + VGAP))
+    done
 }
 
 main() {
-    if [ ! -z $mpvWid ]; then
-        exit
-    else
-        detection.sh
-        pretty.sh
+    source fyrerc.sh
+    source detection.sh
+
+    # Calculate usable screen size (without borders and gaps)
+    SW=$((SW - 2*XGAP - BW))
+    SH=$((SH - TGAP - BGAP - BW))
+
+    ignore
+
+    if [ $windowsOnscreen -le 4 ]; then
+        prettytile
+    fi
+
+    if [ -e $WLFILETEMP ]; then
+        rm $WLFILETEMP
     fi
 }
